Fix get survey response endpoint
A GET operation should not create an entity. Updated the get survey
endpoint to NOT create a response object if it doesn't exist. Instead it
should not return a response as if it was empty without having to create
it.

Also updated HTTP codes in other responses to better match the actual
response state.