Fix survey caching
Updated state for survey to include ID of the last survey that fetch
attempt. This value can be used to determine if the current page has
already attempted to fetch the survey it needs. Previously, if a survey
was already in the state there was no way of knowing if that was the
survey that the page needed.