Re-use userId for login if available
Previously, the login endpoint would generate a new user ID every time a
request was made, which made for a poor user experience in the UI as the
login endpoint is hit on page load. Updated the endpoint to check if a
jwt was included in the request, and will re-use the jwt id if
available. This behavior can be overriden to return a new ID every time
(useful for testing) when running in dev mode, and by setting the
`NEW_USER_ID_PER_JWT` environment variable to a truthy value.